refactor(sdk): move public facade to sdk path
