+++ docs/maintainers/agents/documentation-guidelines.md
@@ -25,7 +25,7 @@
 ## 写作规则
 
 - README 保持安装、能力边界、短示例和文档入口；完整 flag/错误/状态变更契约放 CLI reference。
-- SDK 文档只描述公开 `javdb` package，不把 `internal/` 目录宣称为集成 API。
+- SDK 文档只描述公开 `sdk/` 路径（`package javdb`），不把 `internal/` 目录宣称为集成 API。
 - 架构文档描述当前包边界与运行流，不记录上游逆向过程、签名推导或凭据细节。
 - 长期且有取舍的决策写 ADR；短期实现细节留在代码注释和测试。
 - 命令、配置、环境变量、输出语义、状态变更、构建或测试流程变化时，同步更新相应文档；feature PR
