+++ internal/cli/comments_cmd.go
@@ -6,7 +6,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/FlanChanXwO/javdb-cli/javdb"
+	"github.com/FlanChanXwO/javdb-cli/sdk"
 )
 
 func newCommentsCmd(rf *rootFlags, aio *appIO) *cobra.Command {
