+++ internal/cli/download_cmd.go
@@ -7,7 +7,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/FlanChanXwO/javdb-cli/javdb"
+	"github.com/FlanChanXwO/javdb-cli/sdk"
 )
 
 func newDownloadCmd(rf *rootFlags, aio *appIO) *cobra.Command {
