+++ internal/cli/magnets_cmd.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/FlanChanXwO/javdb-cli/javdb"
+	"github.com/FlanChanXwO/javdb-cli/sdk"
 )
 
 func newMagnetsCmd(rf *rootFlags, aio *appIO) *cobra.Command {
