+++ scripts/test-documentation.sh
@@ -26,8 +26,16 @@ grep -F 'docs/zh-CN/cli-reference.md' "$repo_root/README.zh-CN.md" >/dev/null
 grep -F 'docs/zh-CN/sdk.md' "$repo_root/README.zh-CN.md" >/dev/null
 grep -F 'changelog/README.md' "$repo_root/CHANGELOG.md" >/dev/null
 grep -F 'changelog/README.zh-CN.md' "$repo_root/CHANGELOG.zh-CN.md" >/dev/null
+for document in \
+	"$repo_root/README.md" \
+	"$repo_root/README.zh-CN.md" \
+	"$repo_root/docs/en/sdk.md" \
+	"$repo_root/docs/zh-CN/sdk.md"; do
+	grep -F 'github.com/FlanChanXwO/javdb-cli/sdk' "$document" >/dev/null
+done
 sh "$repo_root/scripts/test-releasenotes.sh"
 grep -F 'internal/javdb/appapi' "$repo_root/AGENTS.md" >/dev/null
+grep -F 'sdk/' "$repo_root/AGENTS.md" >/dev/null
 grep -F '@AGENTS.md' "$repo_root/CLAUDE.md" >/dev/null
 
 if rg -n 'internal/(appapi|httpx|signature)' \
